Fix regular expression syntax
This code is failing because `[0..9]` matches the characters '0', '.',
and '9'. This breaks when the stack name is 'mysql5_5', because it tries
to load 'eydr::mysql5_5_failover' instead of 'eydr::mysql_failover'.

Change the regular expression to be `[0-9]` which correctly matches the
numbers 0 through 9. It will properly split the name of the recipe
and correctly load 'eydr::mysql_failover'