Make eydr chef recipe backwards compatible
The recipe in question is not backwards compatible with the version of
Chef on the EngineYard stack stable-v2.1.1. With this particular version of
Chef the 'file' resrouce does not support the 'content' property.

This uses a workaround to bypass it, by using the template resource and
passing in a single varialbe as the content.